backend-maps/requirements.txt
fastapi
uvicorn
pandas
tqdm
psycopg2-binary
SQLAlchemy
python-dotenv
sqlalchemy-cockroachdb
psycopg2-binary